+++ src/components/Code.mjs
@@ -1,7 +1,7 @@
+import classNameProp from 'class-name-prop'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { stylesCode } from '../styles/stylesCode'
-import { classNameProp } from '../utils/classNameProp'
 import { propTypeChildren } from '../utils/propTypeChildren'
 
 export const Code = React.forwardRef(({ className, ...props }, ref) => (