+++ src/components/LinkText.mjs
@@ -1,7 +1,7 @@
+import classNameProp from 'class-name-prop'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { stylesLinkText } from '../styles/stylesLinkText'
-import { classNameProp } from '../utils/classNameProp'
 import { propTypeChildren } from '../utils/propTypeChildren'
 
 export const LinkText = React.forwardRef(({ className, ...props }, ref) => (