+++ src/utils/index.mjs
@@ -1,2 +1,1 @@
-export { classNameProp } from './classNameProp'
 export { propTypeChildren } from './propTypeChildren'