+++ changelog.md
@@ -9,7 +9,6 @@
 ### Minor
 
 - Added the `stylesSyntaxHighlightingPrism` named export, for styling syntax highlighting HTML produced by [Prism](https://prismjs.com).
-- Added the `classNameProp` named export, for creating a `className` prop value for multiple classes.
 - Added the `propTypeChildren` named export, for use in component `propTypes` to specify a prop (usually, but not necessarily `children`) should have a React node type, and if `.isRequired` is chained, must be populated unless the `dangerouslySetInnerHTML` prop is populated.
 - `Blockquote`, `Button`, `Code`, `Heading`, `Html`, `LinkCard`, `LinkNav`, `LinkText`, `List`, `Nav`, `Para`, `Pre`, `Select`, and `Table` components can now accept a `dangerouslySetInnerHTML` prop.
 - Additional `Nav` component props are now applied to the underlying `nav` element.
@@ -21,7 +20,8 @@
 - Alphabetically sort exports in index files.
 - Fixed setting a `className` prop on `LinkCard` and `LinkNav` components from interfering with the `active` class.
 - The `children` prop of the `Nav` and `Table` components is now marked as required in `propTypes`.
-- The new `classNameProp` and `propTypeChildren` utils have been used in many components.
+- In many components, use the new `propTypeChildren` API.
+- In many components, use `classNameProp` from [`class-name-prop`](https://npm.im/class-name-prop) for creating a `className` prop value for multiple classes.
 
 ## 1.0.1